Add deriveEncoder, deriveDecoder, and deriveCodec methods
